Added player sprite
Mitchell was taking forever to take a shower, so I started working and now he's done so I only had time to get things set up and then add a player placeholder sprite and not add any scripts. We're heading over to the lab soon and hopefully Mitchell won't break the github that I so lovingly set-up for us.